+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Achievement : Add `set_unlock_skin` (TODO)
 - Sprite : Add `.width`, `.height`, and `.subimages` getters to wrapper
 - Callback : Add `add_SO`
+- AttackFlag : Add `new` and `find`
+- AttackInfo : Add support for custom attack flags
 
 ### v0.1.43
 - Survivor : Set `pHmax_base` to `2.8` on init